# apps/core/config/openapi.yml
---
openapi: 3.1.0
components:
  schemas:
    AuthUserModel:
      type: object
      required:
      - id
      - phone
      - firstname
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        phone:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        patronymic:
          type: string
    Authenticator:
      type: string
      enum:
      - TELEGRAM
      - SMS
    ConfirmOtpRequest:
      type: object
      required:
      - phone
      - code
      properties:
        phone:
          type: string
        code:
          type: string
    GoalModel:
      type: object
      required:
      - id
      - userId
      - name
      - durationMonth
      - monthlyInvest
      - target
      - estimatedDate
      - balance
      - completed
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        userId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        durationMonth:
          type: integer
          format: int32
        monthlyInvest:
          type: number
          format: double
        target:
          type: number
          format: double
        estimatedDate:
          $ref: "#/components/schemas/LocalDateTime"
        balance:
          type: number
          format: double
        completed:
          type: boolean
    GoalSaveRequest:
      type: object
      required:
      - name
      - durationMonth
      - monthlyInvest
      - target
      - estimatedDate
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        name:
          type: string
        durationMonth:
          type: integer
          format: int32
        monthlyInvest:
          type: number
          format: double
        target:
          type: number
          format: double
        estimatedDate:
          $ref: "#/components/schemas/LocalDateTime"
    JsonObject:
      type: array
      items:
        type: object
        properties:
          key:
            type: string
          value: {}
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
    ProfileUpdateRequest:
      type: object
      required:
      - firstname
      properties:
        firstname:
          type: string
        lastname:
          type: string
        patronymic:
          type: string
    SaveResult:
      type: object
      required:
      - id
      properties:
        id:
          $ref: "#/components/schemas/UUID"
    SaveSessionDataRequest:
      type: object
    SendOtpRequest:
      type: object
      required:
      - phone
      - authenticator
      properties:
        phone:
          type: string
        authenticator:
          $ref: "#/components/schemas/Authenticator"
    SendOtpResponse:
      type: object
      properties:
        code:
          type: string
    TokenResultModel:
      type: object
      required:
      - token
      properties:
        token:
          type: string
    TransactionModel:
      type: object
      required:
      - id
      - createDate
      - userId
      - amount
      - categoryType
      - date
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        userId:
          $ref: "#/components/schemas/UUID"
        amount:
          type: number
          format: double
        categoryType:
          type: string
        date:
          $ref: "#/components/schemas/LocalDateTime"
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UserModel:
      type: object
      required:
      - id
      - createDate
      - enabled
      - phone
      - firstname
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        enabled:
          type: boolean
        phone:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        patronymic:
          type: string
        uuin:
          type: string
        birthDate:
          $ref: "#/components/schemas/OffsetDateTime"
paths:
  /api/hnu/auth/me:
    delete:
      operationId: deleteMe
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete Me
      tags:
      - Auth Resource
      security:
      - bearer: []
  /api/hnu/auth/my-session:
    put:
      operationId: updateMySession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSessionDataRequest"
        required: true
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update My Session
      tags:
      - Auth Resource
      security:
      - bearer: []
    delete:
      operationId: logout
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Logout
      tags:
      - Auth Resource
      security:
      - bearer: []
  /api/hnu/auth/my-session/{id}:
    put:
      operationId: unlockSessionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
      summary: Unlock Session By Id
      tags:
      - Auth Resource
  /api/hnu/auth/my-user-info:
    post:
      operationId: saveMyUserInfo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Save My User Info
      tags:
      - Auth Resource
      security:
      - bearer: []
    get:
      operationId: getMyUserInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserModel"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get My User Info
      tags:
      - Auth Resource
      security:
      - bearer: []
  /api/hnu/auth/otp:
    post:
      operationId: sendOtpCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          description: Bad Request
      summary: Send Otp Code
      tags:
      - Auth Resource
  /api/hnu/auth/otp-code:
    post:
      operationId: confirmOtpCode
      parameters:
      - name: User-Agent
        in: header
        schema:
          type: string
      - name: X-Real-IP
        in: header
        schema:
          type: string
          default: "-"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmOtpRequest"
        required: true
      responses:
        default:
          description: confirm otp code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResultModel"
        "400":
          description: Bad Request
      summary: Confirm Otp Code
      tags:
      - Auth Resource
  /api/hnu/goals:
    post:
      operationId: saveGoal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
        "400":
          description: Bad Request
      summary: Save Goal
      tags:
      - Goal Resource
    get:
      operationId: getGoals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalModel"
      summary: Get Goals
      tags:
      - Goal Resource
  /api/hnu/goals/user/{userId}:
    get:
      operationId: getUserGoals
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalModel"
      summary: Get User Goals
      tags:
      - Goal Resource
  /api/hnu/goals/{id}:
    delete:
      operationId: removeGoal
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: No Content
      summary: Remove Goal
      tags:
      - Goal Resource
    get:
      operationId: getGoalById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalModel"
      summary: Get Goal By Id
      tags:
      - Goal Resource
  /api/hnu/transactions/user/{userId}:
    get:
      operationId: getUserTransactions
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionModel"
      summary: Get User Transactions
      tags:
      - Transaction Resource
  /api/hnu/users:
    get:
      operationId: getUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserModel"
      summary: Get Users
      tags:
      - User Resource
  /api/hnu/users/{id}:
    get:
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserModel"
      summary: Get User By Id
      tags:
      - User Resource
  /hello:
    get:
      operationId: hello
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
      summary: Hello
      tags:
      - Greeting Resource
  /service/api/fcs/telegram:
    post:
      operationId: onTelegramMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: On Telegram Message
      tags:
      - Telegram Resource
info:
  title: FNU API (development)
  version: 1.0.0